Abstractions taming step 3: the eight GPU uint vocabularies become real enums
GpuComputeAccess, GpuComputeStage, GpuShaderStage, GpuTimingStage are now
[Flags] enum : uint; GpuImageLayout, GpuPixelFormat, GpuSamplerFilter,
GpuComputeBindingKind are plain enum : uint. Member names and numeric
values preserved exactly, so every existing call-site identifier compiles
unchanged and the retype is behavior-identical by construction.

Neutral seam signatures are compiler-checked now: TransitionImageLayout /
MemoryBarrier / TransitionBuffer take (GpuImageLayout, GpuComputeAccess,
GpuComputeStage); PushConstants and GpuPushConstantBinding take
GpuShaderStage; samplers take GpuSamplerFilter; GpuComputeBinding.Kind is
GpuComputeBindingKind; every format parameter is GpuPixelFormat. The
transposition-prone six-uint barrier signatures - a known bug class here -
are dead.

Along for the ride:
- Sharing seam standardized format-first: Upload/Read/Import/
  CreateExportableTarget/CreateExportableStorageImage now take
  (deviceContext, [payload,] format, width, height)
- GpuPixelFormat.FromSurfaceFormat -> GpuPixelFormats.FromSurfaceFormat
  (enums cannot carry methods); the GpuPixelFormat-vs-SurfaceFormat
  divergence rationale is documented on the type
- Backends flow the enums end-to-end through their private mapping
  helpers; casts only at true native boundaries (5 total: VkShaderStage
  passthroughs + VkDescriptorSetLayoutBinding.StageFlags interop struct)
- Puck.Post: 12 one-line mechanical retypes (9 const declarations, 2
  helper params, 1 interop cast), zero logic changes

Verified: solution + Puck.Post build 0 warnings 0 errors; 11-gate GPU
validation battery green on both backends (compute, world vk+dx,
indirect, resample, viewports, pixelate, export, reverse-share, genlock,
mini-action).

+++ src/Puck.Demo/ComputeValidationNode.cs
@@ -125,7 +125,7 @@ internal sealed class ComputeValidationNode : IRenderNode {
                         Binding = 0,
                         DescriptorCount = 1,
                         DescriptorType = DescriptorTypeStorageImage,
-                        StageFlags = GpuShaderStage.Compute,
+                        StageFlags = (uint)GpuShaderStage.Compute,
                     },
                 ],
                 PushConstantSize: 0,
